+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed Migration data with proper active_since and inactive_since values
 - Fix for graph in single lesson view to properly highlight current lesson
 - Added Reports page for generating printable reports containing group and student summaries
+- Added attendance to Group Lesson view and attendance graph for Reports
 
 ## 0.33.0 - Lesson UUIDs & Graphs
 - Added validation for existing tags in an organization